requirements.txt
# Core dependencies with compatible versions
numpy<2.0.0
opencv-python>=4.8.0,<5.0.0
pillow>=10.0.0
matplotlib>=3.7.0
sympy>=1.12

# OCR Libraries
easyocr>=1.7.0
pix2text>=0.3.0

# QR Code processing
qrcode[pil]>=7.4.2
pyzbar>=0.1.9

# Configuration and utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
click>=8.1.7
tqdm>=4.66.1

# Development dependencies
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.9.1
flake8>=6.1.0
isort>=5.12.0
